backend/requirements.txt
fastapi
supabase
python-dotenv
uvicorn
gunicorn
psycopg2-binary
sqlalchemy
passlib[bcrypt]
passlib
email-validator
bcrypt==4.0.1
python-jose[cryptography]
python-multipart
boto3
slowapi
subprocess
fastapi-mail